vtfpp: update face field to uint8_t, update a C enum to match C++

+++ include/vtfpp/ImageConversion.h
@@ -328,7 +328,7 @@ namespace ImageConversion {
 [[nodiscard]] std::vector<std::byte> convertImageDataToFormat(std::span<const std::byte> imageData, ImageFormat oldFormat, ImageFormat newFormat, uint16_t width, uint16_t height);
 
 /// Converts several images from one format to another.
-[[nodiscard]] std::vector<std::byte> convertSeveralImageDataToFormat(std::span<const std::byte> imageData, ImageFormat oldFormat, ImageFormat newFormat, uint8_t mipCount, uint16_t frameCount, uint16_t faceCount, uint16_t width, uint16_t height, uint16_t sliceCount);
+[[nodiscard]] std::vector<std::byte> convertSeveralImageDataToFormat(std::span<const std::byte> imageData, ImageFormat oldFormat, ImageFormat newFormat, uint8_t mipCount, uint16_t frameCount, uint8_t faceCount, uint16_t width, uint16_t height, uint16_t sliceCount);
 
 /// Converts an HDRI into a cubemap. The output image data is in the same image format as the input.
 /// The output images have the following order: front, back, left, right, down, up.
